refactor: 예외처리 관련 기능 Model
